// src/types/type.ts
export interface CrimeReport {
    "Sr no": number;
    "District/City": string;
    loc_name: string;
    fir_no: string;
    case_date: number; 
    Villlage: string;
    Taluka: string;
    "Occurence District": string;
    "Occurence City": string;
    latitude: number;
    longitude: number;
    first_name: string;
    middle_name: string;
    last_name: string;
    age: number;
    Gender: string;
    motive_of_crime: string;
    "IS_DETECTED (Yes/No)": string;
    "Crime Head": string;
    Category: string;
  }
  